pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "bookcart-automation"
version = "1.0.0"
description = "Automated testing framework for BookCart e-commerce application"
authors = [{name = "Test Automation Team"}]
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.40.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "allure-pytest>=2.13.2",
    "pillow>=10.0.0",
    "pydantic>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
